# feat(query): optimize stringsearch like