… Cargo.lock
… src/query/functions/Cargo.toml