I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

optimize like parttern 

0.15s -> 0.10s

```sql
-- before optimize:
mysql> select count() from orders where o_clerk like '%lerk#00000177%';
+---------+
| count() |
+---------+
|   15086 |
+---------+
1 row in set (0.15 sec)
Read 15000000 rows, 329.02 MiB in 0.145 sec., 103.29 million rows/sec., 2.21 GiB/sec.

--after optimize:

mysql> select count() from orders where o_clerk like '%lerk#00000177%';
+---------+
| count() |
+---------+
|   15086 |
+---------+
1 row in set (0.10 sec)
Read 15000000 rows, 329.02 MiB in 0.101 sec., 147.92 million rows/sec., 3.17 GiB/sec.

```

Closes #8458