@TCeason 2022-11-09
In ClickHouse `!= string` cost 0.055 sec, in databend `!= string` cost 0.250 sec, And I retest the int comparison and modify the result in [there](https://github.com/datafuselabs/databend/issues/8458#issuecomment-1308630214).

`!= string` and `int comparison` ck 4x faster than databend.

And in this pr, the ck like function is also 4x faster than databend.

```
databend-arch :) select version();

SELECT version()

Query id: 9e68b4d6-c26c-40b5-9aee-646da7b573b0

┌─version()─┐
│ 22.7.1.1  │
└───────────┘

1 row in set. Elapsed: 0.004 sec. 

databend-arch :) select count() from orders where o_comment != '%s%p%';

SELECT count()
FROM orders
WHERE o_comment != '%s%p%'

Query id: 41b001f2-bf71-4a46-9815-bb3f73460d99

┌──count()─┐
│ 15000000 │
└──────────┘

1 row in set. Elapsed: 0.055 sec. Processed 15.00 million rows, 858.40 MB (274.07 million rows/s., 15.68 GB/s.)

```

```sql
mysql> select count() from orders where o_comment != '%s%p%';
+----------+
| count()  |
+----------+
| 15000000 |
+----------+
1 row in set (0.25 sec)
Read 15000000 rows, 808.13 MiB in 0.250 sec., 60 million rows/sec., 3.16 GiB/sec.

```